Move classproperty into _internal
