+++ src/globus_sdk/client.py
@@ -6,7 +6,7 @@ import typing as t
 import urllib.parse
 
 from globus_sdk import GlobusSDKUsageError, config, exc
-from globus_sdk._classproperty import classproperty
+from globus_sdk._internal.classprop import classproperty
 from globus_sdk._utils import slash_join
 from globus_sdk.authorizers import GlobusAuthorizer
 from globus_sdk.paging import PaginatorTable
